refatorando codigo legado
