+++ 1
@@ -1,0 +1,63 @@
+[199172] Log opened at 2023-11-05 18:22:24.166107
+[199172] [Config] INFO: Trigger value for 'XDEBUG_TRIGGER' not found, falling back to 'XDEBUG_SESSION'
+[199172] [Config] INFO: Trigger value for 'XDEBUG_SESSION' not found, so not activating
+[199172] Log closed at 2023-11-05 18:22:24.481681
+
+[199230] Log opened at 2023-11-05 18:22:42.101670
+[199230] [Config] INFO: Trigger value for 'XDEBUG_TRIGGER' not found, falling back to 'XDEBUG_SESSION'
+[199230] [Config] INFO: Trigger value for 'XDEBUG_SESSION' not found, so not activating
+[199230] Log closed at 2023-11-05 18:22:43.706390
+
+[199315] Log opened at 2023-11-05 18:24:48.207685
+[199315] [Config] INFO: Trigger value for 'XDEBUG_TRIGGER' not found, falling back to 'XDEBUG_SESSION'
+[199315] [Config] INFO: Trigger value for 'XDEBUG_SESSION' not found, so not activating
+[199315] Log closed at 2023-11-05 18:24:49.871716
+
+[199356] Log opened at 2023-11-05 18:25:00.805952
+[199356] [Config] INFO: Trigger value for 'XDEBUG_TRIGGER' not found, falling back to 'XDEBUG_SESSION'
+[199356] [Config] INFO: Trigger value for 'XDEBUG_SESSION' not found, so not activating
+[199356] Log closed at 2023-11-05 18:25:02.902506
+
+[199466] Log opened at 2023-11-05 18:28:34.447313
+[199466] [Config] INFO: Trigger value for 'XDEBUG_TRIGGER' not found, falling back to 'XDEBUG_SESSION'
+[199466] [Config] INFO: Trigger value for 'XDEBUG_SESSION' not found, so not activating
+[199466] Log closed at 2023-11-05 18:28:37.268061
+
+[199541] Log opened at 2023-11-05 18:29:31.050653
+[199541] [Config] INFO: Trigger value for 'XDEBUG_TRIGGER' not found, falling back to 'XDEBUG_SESSION'
+[199541] [Config] INFO: Trigger value for 'XDEBUG_SESSION' not found, so not activating
+[199541] [Config] INFO: Trigger value for 'XDEBUG_TRIGGER' not found, falling back to 'XDEBUG_SESSION'
+[199541] [Config] INFO: Trigger value for 'XDEBUG_SESSION' not found, so not activating
+[199541] [Config] INFO: Trigger value for 'XDEBUG_TRIGGER' not found, falling back to 'XDEBUG_SESSION'
+[199541] [Config] INFO: Trigger value for 'XDEBUG_SESSION' not found, so not activating
+[199541] [Config] INFO: Trigger value for 'XDEBUG_TRIGGER' not found, falling back to 'XDEBUG_SESSION'
+[199541] [Config] INFO: Trigger value for 'XDEBUG_SESSION' not found, so not activating
+[199541] [Config] INFO: Trigger value for 'XDEBUG_TRIGGER' not found, falling back to 'XDEBUG_SESSION'
+[199541] [Config] INFO: Trigger value for 'XDEBUG_SESSION' not found, so not activating
+[199541] [Config] INFO: Trigger value for 'XDEBUG_TRIGGER' not found, falling back to 'XDEBUG_SESSION'
+[199541] [Config] INFO: Trigger value for 'XDEBUG_SESSION' not found, so not activating
+[199541] [Config] INFO: Trigger value for 'XDEBUG_TRIGGER' not found, falling back to 'XDEBUG_SESSION'
+[199541] [Config] INFO: Trigger value for 'XDEBUG_SESSION' not found, so not activating
+[199541] [Config] INFO: Trigger value for 'XDEBUG_TRIGGER' not found, falling back to 'XDEBUG_SESSION'
+[199541] [Config] INFO: Trigger value for 'XDEBUG_SESSION' not found, so not activating
+[199541] Log closed at 2023-11-05 18:29:37.061491
+
+[199615] Log opened at 2023-11-05 18:30:14.400355
+[199615] [Config] INFO: Trigger value for 'XDEBUG_TRIGGER' not found, falling back to 'XDEBUG_SESSION'
+[199615] [Config] INFO: Trigger value for 'XDEBUG_SESSION' not found, so not activating
+[199615] [Config] INFO: Trigger value for 'XDEBUG_TRIGGER' not found, falling back to 'XDEBUG_SESSION'
+[199615] [Config] INFO: Trigger value for 'XDEBUG_SESSION' not found, so not activating
+[199615] [Config] INFO: Trigger value for 'XDEBUG_TRIGGER' not found, falling back to 'XDEBUG_SESSION'
+[199615] [Config] INFO: Trigger value for 'XDEBUG_SESSION' not found, so not activating
+[199615] [Config] INFO: Trigger value for 'XDEBUG_TRIGGER' not found, falling back to 'XDEBUG_SESSION'
+[199615] [Config] INFO: Trigger value for 'XDEBUG_SESSION' not found, so not activating
+[199615] [Config] INFO: Trigger value for 'XDEBUG_TRIGGER' not found, falling back to 'XDEBUG_SESSION'
+[199615] [Config] INFO: Trigger value for 'XDEBUG_SESSION' not found, so not activating
+[199615] [Config] INFO: Trigger value for 'XDEBUG_TRIGGER' not found, falling back to 'XDEBUG_SESSION'
+[199615] [Config] INFO: Trigger value for 'XDEBUG_SESSION' not found, so not activating
+[199615] [Config] INFO: Trigger value for 'XDEBUG_TRIGGER' not found, falling back to 'XDEBUG_SESSION'
+[199615] [Config] INFO: Trigger value for 'XDEBUG_SESSION' not found, so not activating
+[199615] [Config] INFO: Trigger value for 'XDEBUG_TRIGGER' not found, falling back to 'XDEBUG_SESSION'
+[199615] [Config] INFO: Trigger value for 'XDEBUG_SESSION' not found, so not activating
+[199615] Log closed at 2023-11-05 18:32:06.614978
+
